% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event2zeitpunkt.R
\name{event2zeitpunkt}
\alias{event2zeitpunkt}
\title{map between the column EVENTref and the columns zeitpunktref or zp_fabianref
in event2zeitpunkt_table. These columns are different labels of the time points}
\usage{
event2zeitpunkt(
  EVENT,
  returnformat = c("zeitpunktref", "zp_fabianref"),
  event2zeitpunkt_df = progressdatenbankderivate::event2zeitpunkt_table
)
}
\arguments{
\item{EVENT}{vector of characters. They must be present in
event2zeitpunkt_table$EVENTref.}

\item{returnformat}{a character string indicating whether zeitpunktref or
zp_fabianref should be returned.}

\item{event2zeitpunkt_df}{data.table event2zeitpunkt_table (available with
the package).}
}
\value{
a vector of labels of the type zeitpunktref or zp_fabianref
}
\description{
map between the column EVENTref and the columns zeitpunktref or zp_fabianref
in event2zeitpunkt_table. These columns are different labels of the time points
}
\examples{
event2zeitpunkt(EVENT = c(1, 2, 31, 8))
}
