% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event2zeitpunkt_table}
\alias{event2zeitpunkt_table}
\title{A map between different ways to label the time points in the PROGRESS study}
\format{
A data.table with 19 rows and 6 variables
}
\usage{
event2zeitpunkt_table
}
\description{
A data set containing a translation between various ways to label time
points in the PROGRESS study together with their explanation.
}
\keyword{datasets}
