% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProbandInfos.R
\name{extractProbandInfos}
\alias{extractProbandInfos}
\title{get patient characteristics from the table DID_PROBAND}
\usage{
extractProbandInfos(
  DID_PROBAND,
  parameter = c("ZENTRUM", "AGE", "SEX", "WEIGHT", "HEIGHT", "BMI", "COB", "COBM",
    "COBF", "ABKDAYS_BEVOR", "CAPSYS")
)
}
\arguments{
\item{DID_PROBAND}{data.table containing the table DID_PROBAND from the
database of the PROGRESS study}

\item{parameter}{a character vector with the names of the charcteristics
to be extracted.}
}
\value{
a data.table with the PATSTUID and the extracted or computed patient
characteristics.
}
\description{
get patient characteristics from the table DID_PROBAND
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_PROBAND <- readxl::read_excel(excel_fn, "DID_PROBAND")
data.table::setDT(DID_PROBAND)
extractProbandInfos(DID_PROBAND, parameter = "BMI")
}
}
