% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProgressEndpoints.R
\name{extractProgressEndpoints}
\alias{extractProgressEndpoints}
\title{get age and sex from the table DID_PROBAND}
\usage{
extractProgressEndpoints(
  DID_SOFA,
  DID_EP_SERDIS = DID_EP_SERDIS,
  DID_CLIN = DID_CLIN,
  FRM_DIL_LABORWERTE = FRM_DIL_LABORWERTE,
  critMindVerschlechterung = 4,
  sofaEPonly = F
)
}
\arguments{
\item{DID_SOFA}{data.table containing the table DID_SOFA from the database
of the PROGRESS study}

\item{DID_EP_SERDIS}{as above}

\item{DID_CLIN}{as above}

\item{FRM_DIL_LABORWERTE}{as above}

\item{critMindVerschlechterung}{I do not know}

\item{sofaEPonly}{I do not know}
}
\value{
amed list with two elements "detailed" and "compact" containing
data.tables.
}
\description{
get age and sex from the table DID_PROBAND
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_SOFA <- readxl::read_excel(excel_fn, "DID_SOFA")
DID_EP_SERDIS <- readxl::read_excel(excel_fn, "DID_EP_SERDIS")
DID_CLIN <- readxl::read_excel(excel_fn, "DID_CLIN")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
data.table::setDT(DID_SOFA)
data.table::setDT(DID_EP_SERDIS)
data.table::setDT(DID_CLIN)
data.table::setDT(FRM_DIL_LABORWERTE)
extractProgressEndpoints(DID_SOFA, DID_EP_SERDIS, DID_CLIN,
FRM_DIL_LABORWERTE)
}
}
