% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRauchstatus.R
\name{extractRauchstatus}
\alias{extractRauchstatus}
\title{get data about smoking behavior from the table FRM_BAS}
\usage{
extractRauchstatus(
  FRM_BAS,
  rauchvars = c("RAUCHJAHRE", "RAUCHPACK", "RAUCHAKT", "RAUCHL12M")
)
}
\arguments{
\item{FRM_BAS}{data.table containing the table FRM_BAS from the
database of the PROGRESS study}

\item{rauchvars}{a character vector with the names of the variables
to be extracted.}
}
\value{
a data.table with the PATSTUID and the extracted patient
characteristics.
}
\description{
get data about smoking behavior from the table FRM_BAS
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_BAS <- readxl::read_excel(excel_fn, "FRM_BAS")
data.table::setDT(FRM_BAS)
extractRauchstatus(FRM_BAS, rauchvars = "RAUCHL12M")
extractRauchstatus(FRM_BAS, rauchvars = c("RAUCHL12M", "RAUCHJAHRE"))
}
}
