% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4abiot_chrbettl.R
\name{getData4abiot_chrbettl}
\alias{getData4abiot_chrbettl}
\title{Get data on antibiosis before hospital admission (ABIOT), chronic bed
confinement (CHRBETTL).}
\usage{
getData4abiot_chrbettl(FRM_BAS)
}
\arguments{
\item{FRM_BAS}{data.table containing the table FRM_BAS
from the database of the PROGRESS study}
}
\value{
data.table with the ID of the patient (PATSTUID),
time point (EVENT), and the information on antibiosis before hospital
admission (ABIOT), chronic bed confinement (CHRBETTL). Table FRM_BAS
contains values at EVENT == 3 at present.
}
\description{
Get data on antibiosis before hospital admission (ABIOT), chronic bed
confinement (CHRBETTL).
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_BAS <- readxl::read_excel(excel_fn, 'FRM_BAS')
data.table::setDT(FRM_BAS)
res <- getData4abiot_chrbettl(FRM_BAS)
res[]
}
}
