% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4albumin.R
\name{getData4albumin}
\alias{getData4albumin}
\title{Get the albumin levels from the database of the PROGRESS study}
\usage{
getData4albumin(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table so named from
the database of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
albumin level, in the wide format.
}
\description{
Get the albumin levels from the database of the PROGRESS study
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
data.table::setDT(FRM_DIL_LABORWERTE)
toadd_alb <- getData4albumin(FRM_DIL_LABORWERTE)
toadd_alb
}
}
