% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4apo2.min.R
\name{getData4apo2.min}
\alias{getData4apo2.min}
\title{Get data on the minimum of the part pressure of arterial O2.}
\usage{
getData4apo2.min(FRM_O2A)
}
\arguments{
\item{FRM_O2A}{data.table containing the table FRM_O2A from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on the part pressure of arterial O2 (in mmHg), in the wide format.
Small values are bad for the patient.
}
\description{
Get data on the minimum of the part pressure of arterial O2.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_O2A <- readxl::read_excel(excel_fn, 'FRM_O2A')
data.table::setDT(FRM_O2A)
toadd_apo2.min <- getData4apo2.min(FRM_O2A)
toadd_apo2.min
}
}
