% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4art.ph.R
\name{getData4art.ph}
\alias{getData4art.ph}
\title{Get information on the minimum of arterial PH.}
\usage{
getData4art.ph(FRM_B24, FRM_O2A)
}
\arguments{
\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}

\item{FRM_O2A}{data.table containing the table FRM_BEF from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the minimum of the arterial PH, in the wide format.
The minimization is over APH (table FRM_O2A) and APHMIN (table FRM_B24).
Small values are bad for the patient.
}
\description{
Get information on the minimum of arterial PH.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_O2A <- readxl::read_excel(excel_fn, 'FRM_O2A')
data.table::setDT(FRM_B24)
data.table::setDT(FRM_O2A)
toadd_art.ph <- getData4art.ph(FRM_B24, FRM_O2A)
toadd_art.ph
}
}
