% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4beat.R
\name{getData4beat}
\alias{getData4beat}
\title{Get the data on the (mechanical) ventilation.}
\usage{
getData4beat(FRM_BEAT)
}
\arguments{
\item{FRM_BEAT}{data.table containing the table so named from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on the ventilation, in the wide format.
}
\description{
Get the data on the (mechanical) ventilation.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_BEAT <- readxl::read_excel(excel_fn, 'FRM_BEAT')
data.table::setDT(FRM_BEAT)
toadd_beat <- getData4beat(FRM_BEAT)
toadd_beat
}
}
