% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4bemin.R
\name{getData4bemin}
\alias{getData4bemin}
\title{get data on base excess}
\usage{
getData4bemin(FRM_B24)
}
\arguments{
\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain the data on base excess (Basenueberschuss) at
a given time.
}
\description{
get data on base excess
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, "FRM_B24")
data.table::setDT(FRM_B24)
toadd_bemin <- getData4bemin(FRM_B24)
toadd_bemin
}
}
