% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4bun.R
\name{getData4bun}
\alias{getData4bun}
\title{Get information on the maximum of SHARN (urea).}
\usage{
getData4bun(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE
from the database of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the SHARN, in the wide format. BUN is not available
because it is probably obsolete (BUN x 2.143 = Serum Harnstoff (urea)).
}
\description{
Get information on the maximum of SHARN (urea).
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, 'FRM_DIL_LABORWERTE')
data.table::setDT(FRM_DIL_LABORWERTE)
toadd_bun <- getData4bun(FRM_DIL_LABORWERTE)
toadd_bun
}
}
