% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4curb.R
\name{getData4curb}
\alias{getData4curb}
\title{get the CURB65 score from the table DID_CURB65}
\usage{
getData4curb(DID_CURB65)
}
\arguments{
\item{DID_CURB65}{data.table containing the table DID_CURB65 from the database
of the PROGRESS study}
}
\value{
data.table of DID_CURB65 with -1 replaced by NA and added CRB65 column
}
\description{
get the CURB65 score from the table DID_CURB65
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_CURB65 <- readxl::read_excel(excel_fn, 'DID_CURB65')
data.table::setDT(DID_CURB65)
getData4curb(DID_CURB65)
}
}
