% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4diasbp.min.R
\name{getData4diasbp.min}
\alias{getData4diasbp.min}
\title{Get information on the minimum of the diastolic blood pressure. The
minimization is over EVENT within each PATSTUID.}
\usage{
getData4diasbp.min(FRM_RR)
}
\arguments{
\item{FRM_RR}{data.table containing the table FRM_RR from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the maximum of the diastolic blood pressure, in the wide
format
}
\description{
Get information on the minimum of the diastolic blood pressure. The
minimization is over EVENT within each PATSTUID.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_RR <- readxl::read_excel(excel_fn, 'FRM_RR')
data.table::setDT(FRM_RR)
getData4diasbp.min(FRM_RR)
}
}
