% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4diurValue.R
\name{getData4diurValue}
\alias{getData4diurValue}
\title{get data on diuresis}
\usage{
getData4diurValue(FRM_B24)
}
\arguments{
\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain the data on diuresis at a given time.
}
\description{
get data on diuresis
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, "FRM_B24")
data.table::setDT(FRM_B24)
toadd_diurValue <- getData4diurValue(FRM_B24)
toadd_diurValue
}
}
