% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4gcs.R
\name{getData4gcs}
\alias{getData4gcs}
\title{Get the value of the Glasgow Coma Scale}
\usage{
getData4gcs(DID_CLIN)
}
\arguments{
\item{DID_CLIN}{data.table containing the table DID_CLIN from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the Glasgow Coma Scale, in the wide format.
}
\description{
Get the value of the Glasgow Coma Scale
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_CLIN <- readxl::read_excel(excel_fn, 'DID_CLIN')
data.table::setDT(DID_CLIN)
getData4gcs(DID_CLIN)
}
}
