% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4gluk.R
\name{getData4gluk}
\alias{getData4gluk}
\title{get blood glucose data from the table FRM_DIL_LABORWERTE}
\usage{
getData4gluk(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE from the database
of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain blood glucose measurements at a given time.
}
\description{
get blood glucose data from the table FRM_DIL_LABORWERTE
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
data.table::setDT(FRM_DIL_LABORWERTE)
getData4gluk(FRM_DIL_LABORWERTE)
}
}
