% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4haemkrt.R
\name{getData4haemkrt}
\alias{getData4haemkrt}
\title{Get data on haematocrit.}
\usage{
getData4haemkrt(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE
from the database of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on haematocrit, in the wide format.
}
\description{
Get data on haematocrit.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, 'FRM_DIL_LABORWERTE')
data.table::setDT(FRM_DIL_LABORWERTE)
toadd_haemkrt <- getData4haemkrt(FRM_DIL_LABORWERTE)
toadd_haemkrt
}
}
