% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4hfrqMax.R
\name{getData4hfrqMax}
\alias{getData4hfrqMax}
\title{Get information on the maximal heart rate. The maximization is over HFREQ from
the table FRM_BEF, and HFREQMIN, HFREQMAX from the table FRM_B24}
\usage{
getData4hfrqMax(FRM_B24, FRM_BEF)
}
\arguments{
\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}

\item{FRM_BEF}{data.table containing the table FRM_BEF from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the maximal heart rate, in the wide format
}
\description{
Get information on the maximal heart rate. The maximization is over HFREQ from
the table FRM_BEF, and HFREQMIN, HFREQMAX from the table FRM_B24
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
data.table::setDT(FRM_B24)
data.table::setDT(FRM_BEF)
getData4hfrqMax(FRM_B24, FRM_BEF)
}
}
