% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4kate.R
\name{getData4kate}
\alias{getData4kate}
\title{get data on KATECHO}
\usage{
getData4kate(FRM_KAT)
}
\arguments{
\item{FRM_KAT}{data.table containing the table FRM_KAT from the database
of the PROGRESS study}
}
\value{
data.table. Each row corresponds one patient, identified by
the column "patstuid". Columns "kat.atall", "kat.d0", "kat.ab.d1" are
logical and indicate, if KATECHO == 1 at any time, at d0 and after d0.
}
\description{
Values other then 1 are interpreted as 0.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_KAT <- readxl::read_excel(excel_fn, "FRM_KAT")
data.table::setDT(FRM_KAT)
toadd_kate <- getData4kate(FRM_KAT)
toadd_kate
}
}
