% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4leuko.R
\name{getData4leuko}
\alias{getData4leuko}
\title{get the leukocyte count from the table FRM_LABORWERTE}
\usage{
getData4leuko(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE
from the database of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain leukocyte count (dbb, min, max)
at a given time.
}
\description{
The values with COL == "LEUKO" in the table FRM_LABORWERTE are returned
with the suffix "_dbb". The maximum over the values with COL in "LEUKO",
"LEUKOMIN" or "LEUKOMAX" within each combination of patient and event
is returned with the suffix "_max". Similarly, the minimum over these
values is returned with the suffix "_min".
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
data.table::setDT(FRM_DIL_LABORWERTE)
leuko_dat <- getData4leuko(FRM_DIL_LABORWERTE)
leuko_dat
}
}
