% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4map.R
\name{getData4map}
\alias{getData4map}
\title{get the data on the mean arterial pressure}
\usage{
getData4map(DID_CLIN)
}
\arguments{
\item{DID_CLIN}{data.table containing the table DID_CLIN from the database
of the PROGRESS study}
}
\value{
data.table in the wide format with one row per patient, identified
by PATSTUIDs in the column "patstuid" and the values of the mean arterial
pressure at each time.
}
\description{
Extract data with CLIN_PARAM == "MAP-MIN" from the table DID_CLIN and
transform it to the wide format.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_CLIN <- readxl::read_excel(excel_fn, "DID_CLIN")
data.table::setDT(DID_CLIN)
toadd_map <- getData4map(DID_CLIN)
toadd_map
}

}
