% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4multl_inf.R
\name{getData4multl_inf}
\alias{getData4multl_inf}
\title{Get data on the presence of multilobar infiltrates.}
\usage{
getData4multl_inf(FRM_BEF, FRM_VIS)
}
\arguments{
\item{FRM_BEF}{data.table containing the table FRM_BEF from the database
of the PROGRESS study}

\item{FRM_VIS}{data.table containing the table FRM_VIS from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on the presence of multilobar infiltrates, in the wide format.
}
\description{
Get data on the presence of multilobar infiltrates.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
FRM_VIS <- readxl::read_excel(excel_fn, 'FRM_VIS')
data.table::setDT(FRM_BEF)
data.table::setDT(FRM_VIS)
toadd_multl_inf <- getData4multl_inf(FRM_BEF, FRM_VIS)
toadd_multl_inf
}
}
