% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4o2p.min.R
\name{getData4o2p.min}
\alias{getData4o2p.min}
\title{Get data on the oxygen saturation.}
\usage{
getData4o2p.min(FRM_O2P)
}
\arguments{
\item{FRM_O2P}{data.table containing the table FRM_O2P from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on the oxygen saturation, in the wide format.
}
\description{
A minimum of the oxygen saturation levels, if there were several
measurements performed with different methods.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_O2P <- readxl::read_excel(excel_fn, 'FRM_O2P')
data.table::setDT(FRM_O2P)
toadd_o2p.min <- getData4o2p.min(FRM_O2P)
toadd_o2p.min
}
}
