% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4oxi.ind.R
\name{getData4oxi.ind}
\alias{getData4oxi.ind}
\title{get the minimum oxygenation index (Horowitz index) in kPa}
\usage{
getData4oxi.ind(DID_CLIN)
}
\arguments{
\item{DID_CLIN}{data.table containing the table DID_CLIN from the database
of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain the oxygenation index at a given time.
}
\description{
Values from the rows with CLIN_PARAM == "OxygenIndex-MIN" in the table
DID_CLIN are returned.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_CLIN <- readxl::read_excel(excel_fn, "DID_CLIN")
data.table::setDT(DID_CLIN)
toadd_oxi.ind <- getData4oxi.ind(DID_CLIN)
toadd_oxi.ind
}
}
