% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4oxyIndex.R
\name{getData4oxyIndex}
\alias{getData4oxyIndex}
\title{get the minimum oxygenation index (Horowitz index) in mmHg}
\usage{
getData4oxyIndex(DID_OXYGENIND_SINGLE)
}
\arguments{
\item{DID_OXYGENIND_SINGLE}{data.table containing the table so named
from the database of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain the oxygenation index at a given time.
}
\description{
Minima of the values from the table DID_OXYGENIND_SINGLE are returned. The
minimum is computed for each combination of PATSTUID
and EVENT (time point), and is taken over the values of APO2 (table FRM_O2A)
and POXY (table FRM_O2P).
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_OXYGENIND_SINGLE <- readxl::read_excel(excel_fn, "DID_OXYGENIND_SINGLE")
data.table::setDT(DID_OXYGENIND_SINGLE)
toadd_oxyIndex.min <- getData4oxyIndex(DID_OXYGENIND_SINGLE)
toadd_oxyIndex.min
}
}
