% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4pco2.R
\name{getData4pco2}
\alias{getData4pco2}
\title{Get data on the partial pressure of carbon dioxide PCO2MIN.}
\usage{
getData4pco2(FRM_B24)
}
\arguments{
\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on the partial pressure of CO2 (in kPa), in the wide format
(at EVENT == 3 only).
}
\description{
The fucntion gets the value of PCO2MIN and returns its maximum for each
patient and EVENT == 3 (inclusion) in kPa.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
data.table::setDT(FRM_B24)
toadd_pco2.max <- getData4pco2(FRM_B24)
toadd_pco2.max
}
}
