% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4smkern.neutro.R
\name{getData4smkern.neutro}
\alias{getData4smkern.neutro}
\title{get the count of segmented neutrophils}
\usage{
getData4smkern.neutro(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE
from the database of the PROGRESS study}
}
\value{
data.table in the wide format. Each row corresponds to one patient
and the columns contain the count of segmented neutrophils at a given time.
}
\description{
The conunt is per 1 microliter.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
data.table::setDT(FRM_DIL_LABORWERTE)
toadd_smkern.neutro <- getData4smkern.neutro(FRM_DIL_LABORWERTE)
toadd_smkern.neutro
}
}
