% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4SNAT.R
\name{getData4SNAT}
\alias{getData4SNAT}
\title{Get information on SNAT.}
\usage{
getData4SNAT(FRM_DIL_LABORWERTE)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE
from the database of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on SNAT, in the wide format.
}
\description{
Get information on SNAT.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, 'FRM_DIL_LABORWERTE')
data.table::setDT(FRM_DIL_LABORWERTE)
toadd_snat <- getData4SNAT(FRM_DIL_LABORWERTE)
toadd_snat
}
}
