% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4temp.R
\name{getData4temp}
\alias{getData4temp}
\title{Get information on the body temperature.}
\usage{
getData4temp(FRM_BEF, FRM_B24)
}
\arguments{
\item{FRM_BEF}{data.table containing the table FRM_BEF from the database
of the PROGRESS study}

\item{FRM_B24}{data.table containing the table FRM_B24 from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on the body temperature, in the wide format
}
\description{
Get information on the body temperature.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
data.table::setDT(FRM_B24)
data.table::setDT(FRM_BEF)
toadd_temp <- getData4temp(FRM_B24, FRM_BEF)
toadd_temp
}
}
