% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4thrombo.R
\name{getData4thrombo}
\alias{getData4thrombo}
\title{get thrombocyte count from the database of the PROGRESS study}
\usage{
getData4thrombo(FRM_DIL_LABORWERTE, DID_CLIN)
}
\arguments{
\item{FRM_DIL_LABORWERTE}{data.table containing the table FRM_DIL_LABORWERTE from the database
of the PROGRESS study}

\item{DID_CLIN}{as above}
}
\value{
data.table
}
\description{
get thrombocyte count from the database of the PROGRESS study
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
DID_CLIN <- readxl::read_excel(excel_fn, "DID_CLIN")
data.table::setDT(FRM_DIL_LABORWERTE)
data.table::setDT(DID_CLIN)
getData4thrombo(FRM_DIL_LABORWERTE, DID_CLIN)
}

}
