% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData4verwirrt.R
\name{getData4verwirrt}
\alias{getData4verwirrt}
\title{get information on whether the patient is confused (verwirrt) from the table FRM_BEF}
\usage{
getData4verwirrt(FRM_BEF)
}
\arguments{
\item{FRM_BEF}{data.table containing the table FRM_BEF from the database
of the PROGRESS study}
}
\value{
data.table with the ID of the patient (patstuid), and the
information on whether the patient is confused, in the wide format
}
\description{
get information on whether the patient is confused (verwirrt) from the table FRM_BEF
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
data.table::setDT(FRM_BEF)
toadd_verwirrt <- getData4verwirrt(FRM_BEF)
toadd_verwirrt[]
}
}
