% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HalmScore.R
\name{HalmScore}
\alias{HalmScore}
\title{Compute the Halm score.}
\usage{
HalmScore(
  FRM_B24,
  FRM_BEF,
  FRM_RR,
  FRM_O2A,
  FRM_O2P,
  FRM_BEAT,
  DID_CLIN,
  zp_fabian = "auf_in_d-1_in_d0"
)
}
\arguments{
\item{FRM_B24}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BEF}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_RR}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_O2A}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_O2P}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BEAT}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{DID_CLIN}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{zp_fabian}{vector of characters. They must be present in
event2zeitpunkt_table$zp_fabianref. Currently, only
zp_fabian = "auf_in_d-1_in_d0" is possible.}
}
\value{
a named list with components: input and out. input is a data.table
in the wide format (one row per patient), containing the data used for
computing the Halm score. out is a data.table with one row
corresponding to one patient, identified by the
PATSTUID. The column halm contains the value of Halm. The score is non-NA,
if more than 50% of its 6 subscores, i.e., at least 4 subscores are non-NA.
}
\description{
Compute the Halm score.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
FRM_RR <- readxl::read_excel(excel_fn, 'FRM_RR')
FRM_O2A <- readxl::read_excel(excel_fn, 'FRM_O2A')
FRM_O2P <- readxl::read_excel(excel_fn, 'FRM_O2P')
FRM_BEAT <- readxl::read_excel(excel_fn, 'FRM_BEAT')
DID_CLIN <- readxl::read_excel(excel_fn, 'DID_CLIN')
data.table::setDT(FRM_B24)
data.table::setDT(FRM_BEF)
data.table::setDT(FRM_RR)
data.table::setDT(FRM_O2A)
data.table::setDT(FRM_O2P)
data.table::setDT(FRM_BEAT)
data.table::setDT(DID_CLIN)
erg <- HalmScore(FRM_B24, FRM_BEF, FRM_RR, FRM_O2A, FRM_O2P, FRM_BEAT,
DID_CLIN, zp_fabian = "auf_in_d-1_in_d0")
erg
}
}
