% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HalmScore_simple.R
\name{HalmScore_simple}
\alias{HalmScore_simple}
\title{Compute the Halm score.}
\usage{
HalmScore_simple(
  hfrq.max,
  sysbp.min,
  afrq.max,
  o2p.min,
  apo2.min,
  bea,
  sauerst,
  temp.max,
  verwirrt,
  gcs
)
}
\arguments{
\item{hfrq.max}{Pulse in beats/min}

\item{sysbp.min}{Systolic blood pressure in mmHg}

\item{afrq.max}{Respiratory rate in breaths/min}

\item{o2p.min}{Minimum oxygen saturation in percent}

\item{apo2.min}{Partial pressure of oxygen in mmHg}

\item{bea}{mechanical ventilation, yes 1, no 0.}

\item{sauerst}{oxygen therapy, yes 1, no 0.}

\item{temp.max}{Maximal temperature in Celsius}

\item{verwirrt}{Altered mental status (confusion), yes 1, no 0.}

\item{gcs}{Glasgow Coma Scale}
}
\value{
The value of the Halm score.
}
\description{
Compute the Halm score.
}
\examples{
halm <- HalmScore_simple(
hfrq.max = 104,
sysbp.min = 95,
afrq.max = 19,
o2p.min = 90,
apo2.min = 50,
bea = 1,
sauerst = 0,
temp.max = 38,
verwirrt = 0,
gcs = 15)
halm
}
