% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.fct.R
\name{psi.fct}
\alias{psi.fct}
\title{Compute the PSI (pneumonia severity index).}
\usage{
psi.fct(
  DID_PROBAND,
  FRM_BAS,
  FRM_BEF,
  FRM_B24,
  FRM_RR,
  FRM_O2A,
  FRM_DIL_LABORWERTE,
  FRM_VIS,
  zp_fabian = "d0",
  event2zeitpunkt_df = progressdatenbankderivate::event2zeitpunkt_table
)
}
\arguments{
\item{DID_PROBAND}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BAS}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BEF}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_B24}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_RR}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_O2A}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_DIL_LABORWERTE}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_VIS}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{zp_fabian}{vector of characters. They must be present in
event2zeitpunkt_table$zp_fabianref.}

\item{event2zeitpunkt_df}{data.table event2zeitpunkt_table (available with
the package).}
}
\value{
data.table with the ID of the patient (patstuid), and the
data on pleural effusion (PLEUERGUSS from FRM_BEF, FRM_VIS),
in the wide format.
}
\description{
Compute the PSI (pneumonia severity index).
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_PROBAND <- readxl::read_excel(excel_fn, 'DID_PROBAND')
FRM_BAS <- readxl::read_excel(excel_fn, 'FRM_BAS')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_RR <- readxl::read_excel(excel_fn, 'FRM_RR')
FRM_O2A <- readxl::read_excel(excel_fn, 'FRM_O2A')
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
FRM_VIS <- readxl::read_excel(excel_fn, 'FRM_VIS')
data.table::setDT(DID_PROBAND)
data.table::setDT(FRM_BAS)
data.table::setDT(FRM_BEF)
data.table::setDT(FRM_B24)
data.table::setDT(FRM_RR)
data.table::setDT(FRM_O2A)
data.table::setDT(FRM_DIL_LABORWERTE)
data.table::setDT(FRM_VIS)
erg <- psi.fct(DID_PROBAND, FRM_BAS, FRM_BEF, FRM_B24, FRM_RR, FRM_O2A,
FRM_DIL_LABORWERTE,FRM_VIS)
erg
}
}
