% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_simple.R
\name{psi_simple}
\alias{psi_simple}
\title{Calculate the PSI score}
\usage{
psi_simple(
  age,
  verwirrt,
  hfrq.max,
  afrq.max,
  sysbp.min,
  temp.min,
  temp.max,
  tumor,
  herz,
  cerebro,
  renal,
  liver,
  gender,
  nurse.home,
  art.ph.min,
  bun,
  snat,
  gluk,
  haemkrt,
  apo2.min,
  pleu_erg
)
}
\arguments{
\item{age}{Age}

\item{verwirrt}{Altered mental status (confusion), yes 1, no 0}

\item{hfrq.max}{Pulse in beats/min}

\item{afrq.max}{Respiratory rate in breaths/min}

\item{sysbp.min}{Systolic blood pressure in mmHg}

\item{temp.min}{Minimal temperature in Celsius}

\item{temp.max}{Maximal temperature in Celsius}

\item{tumor}{Neoplastic disease, yes 1, no 0}

\item{herz}{Congestive heart failure history, yes 1, no 0}

\item{cerebro}{Cerebrovascular disease history}

\item{renal}{Renal disease history}

\item{liver}{Liver disease history}

\item{gender}{sex, male 0, female 1}

\item{nurse.home}{Nursing home resident, yes 1, no 0}

\item{art.ph.min}{pH}

\item{bun}{BUN in mmol/L}

\item{snat}{Snat (Sodium) in mmol/L}

\item{gluk}{Glucose in mmol/L}

\item{haemkrt}{Hematocrit in percent}

\item{apo2.min}{Partial pressure of oxygen in mmHg}

\item{pleu_erg}{Pleural effusion on x-ray}
}
\value{
The value of the PSI score (positive integer).
}
\description{
Calculate the PSI score
}
\examples{
psi <- psi_simple(
  age = 68,
  verwirrt = 1,
  hfrq.max = 130,
  afrq.max = 35,
  sysbp.min = 80,
  temp.min = 34,
  temp.max = 35,
  tumor = 1,
  herz = 1,
  cerebro = 1,
  renal = 1,
  liver = 1,
  gender = 1,
  nurse.home = 1,
  art.ph.min = 7.2,
  bun = 10,
  snat = 120,
  gluk = 16,
  haemkrt = 0.2,
  apo2.min = 50,
  pleu_erg = 1)
  psi
}
