% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickSOFA.R
\name{quickSOFA}
\alias{quickSOFA}
\title{Compute the quickSOFA score.}
\usage{
quickSOFA(FRM_RR, FRM_B24, FRM_BEF, DID_CLIN, zp_fabian = "auf_in_d0")
}
\arguments{
\item{FRM_RR}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_B24}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BEF}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{DID_CLIN}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{zp_fabian}{vector of characters. They must be present in
event2zeitpunkt_table$zp_fabianref. Currently, only zp_fabian = "auf_in_d0"
is possible.}
}
\value{
a named list with components: input and out. input is a data.table
in the wide format (one row per patient), containing the data used for
computing the quickSOFA score. out is a data.table with one row
corresponding to one patient, identified by the
PATSTUID. The column res contains the value of quickSOFA.
}
\description{
Compute the quickSOFA score.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
FRM_RR <- readxl::read_excel(excel_fn, 'FRM_RR')
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
DID_CLIN <- readxl::read_excel(excel_fn, 'DID_CLIN')
data.table::setDT(FRM_RR)
data.table::setDT(FRM_B24)
data.table::setDT(FRM_BEF)
data.table::setDT(DID_CLIN)
erg_d0 <- quickSOFA(FRM_RR, FRM_B24, FRM_BEF,
DID_CLIN, zp_fabian = "auf_in_d0")
erg_d0
}
}
