% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs.fct.R
\name{sirs.fct}
\alias{sirs.fct}
\title{Compute the SIRS score.}
\usage{
sirs.fct(
  DID_PROBAND,
  FRM_BAS,
  FRM_BEF,
  FRM_B24,
  FRM_DIL_LABORWERTE,
  DID_CLIN,
  FRM_RR,
  FRM_KAT,
  zp_fabian = "d0",
  event2zeitpunkt_df = progressdatenbankderivate::event2zeitpunkt_table
)
}
\arguments{
\item{DID_PROBAND}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BAS}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_BEF}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_B24}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_DIL_LABORWERTE}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{DID_CLIN}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_RR}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{FRM_KAT}{data.table containing the table with the same name from
the database of the PROGRESS study}

\item{zp_fabian}{vector of characters. They must be present in
event2zeitpunkt_table$zp_fabianref.}

\item{event2zeitpunkt_df}{data.table event2zeitpunkt_table (available with
the package).}
}
\value{
a named list with components: input, input2 and out. out is
a data.table with one row corresponding to a combination of PATSTUID
(patient) and "EVENT" (time point).
}
\description{
Compute the SIRS score.
}
\examples{
\dontrun{
excel_fn <- paste0("/net/ifs1/san_projekte/projekte/",
"PROGRESS/Datenmanagement/Data_freezes/",
"20190320/PROGRESS-freeze_201903_01.xlsx")
DID_PROBAND <- readxl::read_excel(excel_fn, 'DID_PROBAND')
FRM_BAS <- readxl::read_excel(excel_fn, 'FRM_BAS')
FRM_BEF <- readxl::read_excel(excel_fn, 'FRM_BEF')
FRM_B24 <- readxl::read_excel(excel_fn, 'FRM_B24')
FRM_DIL_LABORWERTE <- readxl::read_excel(excel_fn, "FRM_DIL_LABORWERTE")
DID_CLIN <- readxl::read_excel(excel_fn, 'DID_CLIN')
FRM_RR <- readxl::read_excel(excel_fn, 'FRM_RR')
FRM_KAT <- readxl::read_excel(excel_fn, 'FRM_KAT')
data.table::setDT(DID_PROBAND)
data.table::setDT(FRM_BAS)
data.table::setDT(FRM_BEF)
data.table::setDT(FRM_B24)
data.table::setDT(FRM_DIL_LABORWERTE)
data.table::setDT(DID_CLIN)
data.table::setDT(FRM_RR)
data.table::setDT(FRM_KAT)
erg_d0 <- sirs.fct(DID_PROBAND, FRM_BAS, FRM_BEF, FRM_B24,
FRM_DIL_LABORWERTE, DID_CLIN, FRM_RR, FRM_KAT, zp_fabian = "d0")
erg_d0
erg_d1 <- sirs.fct(DID_PROBAND, FRM_BAS, FRM_BEF, FRM_B24,
FRM_DIL_LABORWERTE, DID_CLIN, FRM_RR, FRM_KAT, zp_fabian = "d1")
erg_d1
}
}
