% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs_simple.R
\name{sirs_simple}
\alias{sirs_simple}
\title{Calulate the SIRS score}
\usage{
sirs_simple(
  temp.min,
  temp.max,
  hfrq.max,
  afrq.max,
  pco2,
  leuko_min,
  leuko_max,
  stkern.neutro,
  smkern.neutro,
  verwirrt,
  thrombo_min,
  thrombo.daybefore,
  oxi.ind,
  chr.lunge,
  diur,
  gewicht,
  bemin,
  sysbp.min,
  map,
  kate
)
}
\arguments{
\item{temp.min}{Minimal temperature in Celsius}

\item{temp.max}{Maximal temperature in Celsius}

\item{hfrq.max}{Pulse in beats/min}

\item{afrq.max}{Respiratory rate in breaths/min}

\item{pco2}{Arterial partial pressure of carbon dioxide in kPa}

\item{leuko_min}{Minimal WBC (white blood cell) count per cubic cm}

\item{leuko_max}{Maximal WBC (white blood cell) count per cubic cm}

\item{stkern.neutro}{Banded neutrophils count per microliter}

\item{smkern.neutro}{Segmented neutrophils count per microliter}

\item{verwirrt}{Altered mental status (confusion), yes 1, no 0}

\item{thrombo_min}{Minimal thrombocyte count per microlitre}

\item{thrombo.daybefore}{Minimal thrombocyte count per microlitre
on the previous day}

\item{oxi.ind}{Minimal value of the oxygenation (Horowitz) index in kPa}

\item{chr.lunge}{Chronic lung disease, yes 1, no 0}

\item{diur}{Diuresis in ml per day}

\item{gewicht}{Weight in kg}

\item{bemin}{Base excess}

\item{sysbp.min}{Systolic blood pressure in mmHg}

\item{map}{Mean arterial pressure in kPa}

\item{kate}{Treatment using catecholamines, zes 1, no 0}
}
\value{
A list with two compontens. The first, "infec.septic.servsept"
is a scalar or vector with three possible values representing the SIRS
criteria which the patient meets: 1: infected, 2: septic, 3: severe septic.
 The second, "shock" is a logical indicator if the septic shock criteria are
 met.
}
\description{
Calulate the SIRS score
}
\examples{
sirs <- sirs_simple(
  temp.min = 34,
  temp.max = 35,
  hfrq.max = 104,
  afrq.max = 20,
  pco2 = 4.1,
  leuko_min = 7800,
  leuko_max = 7900,
  stkern.neutro = 395,
  smkern.neutro = 4424,
  verwirrt = 1,
  thrombo_min = 216000,
  thrombo.daybefore = 205000,
  oxi.ind = 43.81,
  chr.lunge = 1,
  diur = 3120,
  gewicht = 70,
  bemin = 3.4,
  sysbp.min = 95,
  map = 10.44,
  kate = FALSE)
  sirs
}
